Fixed the aliases

+++ Channels/Alias.cs
@@ -6,31 +6,31 @@ using System.Threading.Tasks;
 
 namespace WGP.CHANNELS
 {
-    public class DoubleChannel : Channel<float> { }
+    public class DoubleChannel : Channel<double> { }
 
-    public class DoubleMasterChannel : MasterChannel<float> { }
+    public class DoubleMasterChannel : MasterChannel<double> { }
 
     public class FloatChannel : Channel<float> { }
 
     public class FloatMasterChannel : MasterChannel<float> { }
 
-    public class IntChannel : Channel<float> { }
+    public class IntChannel : Channel<int> { }
 
-    public class IntMasterChannel : MasterChannel<float> { }
+    public class IntMasterChannel : MasterChannel<int> { }
 
-    public class LongChannel : Channel<float> { }
+    public class LongChannel : Channel<long> { }
 
-    public class LongMasterChannel : MasterChannel<float> { }
+    public class LongMasterChannel : MasterChannel<long> { }
 
     public class TransformChannel : Channel<SFML.Graphics.Transform> { }
 
     public class TransformMasterChannel : MasterChannel<SFML.Graphics.Transform> { }
 
-    public class UIntChannel : Channel<float> { }
+    public class UIntChannel : Channel<uint> { }
 
-    public class UIntMasterChannel : MasterChannel<float> { }
+    public class UIntMasterChannel : MasterChannel<uint> { }
 
-    public class ULongChannel : Channel<float> { }
+    public class ULongChannel : Channel<ulong> { }
 
-    public class ULongMasterChannel : MasterChannel<float> { }
+    public class ULongMasterChannel : MasterChannel<ulong> { }
 }
